slightly restructure files
