+++ packages/core/index.ts
@@ -1,0 +1,3 @@
+// This file serves as an entry point for the package
+export { Lenis as default } from './src/lenis'
+export * from './src/types'
