+++ packages/react/index.ts
@@ -1,0 +1,8 @@
+// This file serves as an entry point for the package
+export {
+  ReactLenis as Lenis,
+  ReactLenis,
+  ReactLenis as default,
+} from './src/provider'
+export * from './src/types'
+export { useLenis } from './src/use-lenis'
