+++ packages/snap/index.ts
@@ -1,0 +1,3 @@
+// This file serves as an entry point for the package
+export { Snap as default } from './src/snap'
+export * from './src/types'
