+++ packages/vue/index.ts
@@ -1,0 +1,8 @@
+// This file serves as an entry point for the package
+export {
+  VueLenis as Lenis,
+  VueLenis,
+  vueLenisPlugin as default,
+} from './src/provider'
+export * from './src/types'
+export { useLenis } from './src/use-lenis'
