+++ tsup.config.ts
@@ -36,42 +36,39 @@ function makeBuildOptions<F extends Format>(
 // Builds
 export const coreESMOptions = makeBuildOptions(
   'lenis',
-  'packages/core/src/index.ts',
+  'packages/core/index.ts',
   'esm'
 )
 const coreIIFEOptions = makeBuildOptions(
   'lenis',
-  'packages/core/src/index.ts',
+  'packages/core/index.ts',
   'iife'
 )
 const coreCSSOptions = makeBuildOptions(
   'lenis',
-  'packages/core/src/lenis.css',
+  'packages/core/lenis.css',
   undefined,
-  { dts: false, sourcemap: false }
+  { dts: false, sourcemap: false, minify: true }
 )
 
 const snapESMOptions = makeBuildOptions(
   'lenis-snap',
-  'packages/snap/src/index.ts',
+  'packages/snap/index.ts',
   'esm'
 )
 const snapIIFEOptions = makeBuildOptions(
   'lenis-snap',
-  'packages/snap/src/index.ts',
+  'packages/snap/index.ts',
   'iife'
 )
 
 const reactOptions = makeBuildOptions(
   'lenis-react',
-  'packages/react/src/index.tsx',
-  'esm'
+  'packages/react/index.ts',
+  'esm',
+  { banner: { js: '"use client";' } }
 )
-const vueOptions = makeBuildOptions(
-  'lenis-vue',
-  'packages/vue/src/index.ts',
-  'esm'
-)
+const vueOptions = makeBuildOptions('lenis-vue', 'packages/vue/index.ts', 'esm')
 
 export default defineConfig(() => {
   console.log(`\x1b[31mLNS\x1b[0m\x1b[1m Building all packages\x1b[0m\n`)
